+++ uefi/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added `proto::hii::config_str::ConfigurationString`.
 - Added `proto::acpi::AcpiTable`.
 - Added `proto::hii::database::HiiDatabase`.
+- Added `proto::hii::config_str::MultiConfigurationStringIter`.
 
 ## Changed
 - **Breaking:** `boot::stall` now take `core::time::Duration` instead of `usize`.